Time: 11 ms (48.72%) | Memory: 45.1 MB (89.03%) - LeetSync

+++ 1885-count-number-of-homogenous-substrings/count-number-of-homogenous-substrings.java
@@ -7,13 +7,14 @@ class Solution {
             if(s.charAt(i)==s.charAt(i+1)){
                 continiousCount++;
             }else{
-                long groupSum = (long)continiousCount * (continiousCount + 1) / 2;
-                count = (int)((count + groupSum) % mod);
-                continiousCount = 1;
+                long groupSum=(long)continiousCount*(long)(continiousCount+1)/2;
+                count=(int)(((long)count+groupSum)%mod);
+                continiousCount=1;
             }
         }
-        long groupSum = (long)continiousCount * (continiousCount + 1) / 2;
-        count = (int)((count + groupSum) % mod);
+        long groupSum=(long)continiousCount*(long)(continiousCount+1)/2;
+        count=(int)(((long)count+groupSum)%mod);
+        
         return count;
     }
 }
